+++ ext/himg/src/options.rs
@@ -2,11 +2,27 @@ use crate::image_size::ImageSize;
 
 use blitz_traits::ColorScheme;
 
-#[derive(Clone, Copy)]
+#[derive(Clone)]
 pub struct Options {
     pub image_size: ImageSize,
     pub color_scheme: ColorScheme,
-    pub allow_net_requests: bool,
+    pub base_url: Option<String>,
     pub truncate: bool,
     pub verbose: bool,
+}
+
+impl Options {
+    pub fn default() -> Self {
+        Self {
+            image_size: ImageSize {
+                width: 720,
+                height: 405,
+                hidpi_scale: 1.0,
+            },
+            truncate:  true,
+            verbose:  false,
+            base_url: None,
+            color_scheme: ColorScheme::Light,
+        }
+    }
 }