+++ lib/himg.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require_relative "himg/version"
+require_relative "himg/base_url"
 require "himg/railtie" if defined?(Rails::Railtie)
 
 # Attempt to load a versioned extension based on the Ruby version.
@@ -19,7 +20,7 @@ module Himg
   RENDER_OPTIONS = %i[width height truncate verbose].freeze
   class Error < StandardError; end
 
-  def self.render(html, width: 720, height: 405, truncate: true, verbose: false)
-    render_to_string(html, "width" => width.to_i, "height" => height.to_i, "truncate" => truncate, "verbose" => verbose)
+  def self.render(html, width: 720, height: 405, truncate: true, verbose: false, base_url: nil)
+    render_to_string(html, "width" => width.to_i, "height" => height.to_i, "truncate" => truncate, "verbose" => verbose, "base_url" => BaseUrl.new(base_url).to_s)
   end
 end